requirements.txt
Django==4.1.3
psycopg2
djangorestframework
django-rest-authtoken
django-environ
pytest
pytest-django
matplotlib
django-cors-headers
